+++ CHANGELOG.md
@@ -48,7 +48,7 @@ To learn more about active deprecations, we recommend checking [GitHub Discussio
 
 - **CPU/Memory scaler**: Add support for scale to zero if there are multiple triggers([#4269](https://github.com/kedacore/keda/issues/4269))
 - **Redis Scalers**: Allow scaling using redis stream length ([#4277](https://github.com/kedacore/keda/issues/4277))
-- **Redis Scalers**: Allow scaling using consumer group lag ([PR number](issue))
+- **Redis Scalers**: Allow scaling using consumer group lag ([#4592](https://github.com/kedacore/keda/issues/4592))
 - **General:** Introduce new Solr Scaler ([#4234](https://github.com/kedacore/keda/issues/4234))
 
 ### Improvements